Move constants for testing to own file
