+++ packages/server/src/test/rps_test_data.ts
@@ -27,7 +27,7 @@ import {
   PARTICIPANT_1_PRIVATE_KEY,
   PARTICIPANTS,
   STAKE,
-} from '../constants';
+} from '../test-constants';
 import { default_channel } from './test_data';
 
 // Commitment Constructors
